Improve long long compatibility among different host machines.
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@9524 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/src/z80/gen.c
@@ -4741,7 +4741,7 @@ genRet (const iCode *ic)
       regalloc_dry_run_cost += 8;
       do
         {
-          emit2 ("ld (hl), !immedbyte", lit & 0xff);
+          emit2 ("ld (hl), !immedbyte", (unsigned long) (lit & 0xff));
           regalloc_dry_run_cost += 2;
           lit >>= 8;
           if (size > 1)
@@ -6958,7 +6958,7 @@ gencjneshort (operand * left, operand * right, symbol * lbl, const iCode *ic)
   if (AOP_TYPE (right) == AOP_LIT)
     {
       lit = ullFromVal (AOP (right)->aopu.aop_lit);
-      if (lit == 0)
+      if (lit == 0ull)
         {
           cheapMove (ASMOP_A, 0, AOP (left), offset);
           if (size > 1)
